pyproject.toml
[tool.poetry]
name = "histman"
version = "0.1.0"
description = "ZSH History Manager"
authors = ["Caesar Kabalan <caesar.kabalan@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
click = "^7.1.1"
python-gnupg = "^0.4.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
